+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 0.18.0
 
+* Add possibility to set Java System Properties for UO and TO via CR yaml
 * Make it possible to configure PodManagementPolicy for StatefulSets
 
 ## 0.17.0